Use Hiera data for spec tests
This commit allows us to use our Hiera data that we were previously only
using with Vagrant as a fixture for our spec tests _and_ as
configuration for Vagrant.

Note that there is a caveat when attempting to re-run specs after
launching the Vagrant environment. It has been ticketed as
https://github.com/rji/puppet-waylon/issues/1.