Disable management of resources by rbenv
The rbenv puppet module installs a bunch of packages that we already
manage, and the waylon module depends on rbenv.

This change sets the manage_deps parameter on it to false. Note that the
manage_deps parameter isn't merged in yet and is in
https://github.com/justindowning/puppet-rbenv/pull/57, so this won't
work with the currently released puppet-rbenv.